stress-ng.h: Fix incorrect type in stress_long_double_put
Make argument long double, fix warning:

Fixes warning:
./stress-ng.h:3868:30: warning: implicit conversion increases floating-point
precision: 'const double' to 'long double' [-Wdouble-promotion]
        g_put_val.long_double_val = a;

+++ stress-ng.h
@@ -3863,7 +3863,7 @@ static inline void ALWAYS_INLINE stress_double_put(const double a)
  *  stress_long_double_put()
  *	stash a double value
  */
-static inline void ALWAYS_INLINE stress_long_double_put(const double a)
+static inline void ALWAYS_INLINE stress_long_double_put(const long double a)
 {
 	g_put_val.long_double_val = a;
 }
